ENH: Add new MLP method "FlexibleMLP" which does not relay on external parametrization
The new method uses input and output energy per proton pair and derives 
an 
approximate parametrization of the kinematic scattering power term from 
it. 
The new method is currently only available in DistanceDrivenBinning, not 
in BackProjection. 

The new MLP needs an Init() method with a modified signature, see 
pctMostLikelyPathFunction.h.

Remove obsolete and commented code in several places. 

Implement CanBeVectorised flag for all MLP classes to decided how an 
Evaluate loop needs to be set up.